some changes

+++ usuarios/views.py
@@ -8,7 +8,7 @@ from .models import *
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.urls import reverse_lazy
-from simulados.models import Simulado
+from simulados.models import *
 
 #cordenadores
 class LoginPageView(TemplateView):
@@ -46,11 +46,9 @@ class LoginPageView(TemplateView):
             email = self.request.user.username
             #verificando se o usuario é coordenador
             if Coordenador.objects.filter(email=email).exists():
-                print("coordenador")
                 return redirect('home_coordenador')
             
             else:
-                print("professor")
                 return redirect('home_professor')
         return render(self.request, self.template_name)
     
@@ -136,7 +134,6 @@ class CadastroCoordenadorGeral(SecretarioView, TemplateView):
             return render(self.request, self.template_name, context)
         
         #criando user
-        print("Username: ", username)
         user = User.objects.create_user(
             username=username, 
             email=email, 
@@ -236,8 +233,15 @@ class CoordenadorGeralCreateCoordenadorView(CoordenadorGeralView, TemplateView):
         return redirect('home_coordenador_geral')
 
 ############# COORDENADOR ################
-class CoordenadorView(LoginRequiredMixin):
+class CoordenadorBaseView(LoginRequiredMixin):
     login_url = reverse_lazy('login')
+
+    #criando redirect para deslogar caso o usuario não seja coordenador
+    def dispatch(self, request, *args, **kwargs):
+        if not Coordenador.objects.filter(user=request.user).exists():
+            logout(request)
+            return redirect('login')
+        return super().dispatch(request, *args, **kwargs)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -251,13 +255,13 @@ class CoordenadorView(LoginRequiredMixin):
         turmas = Turma.objects.filter(escola=escola)
         context['turmas'] = turmas
         context['escola'] = escola
-
         return context
-class CoordenadorDashboard(CoordenadorView, TemplateView):
+    
+class DashboardCoordenador(CoordenadorBaseView, TemplateView):
     template_name = 'usuarios/coordenadores/dashboard.html'
 
-class CoordenadorListaProfessoresView(CoordenadorView, ListView):
-    template_name = 'usuarios/coordenadores/lista_professores.html'
+class ListaProfessoresView(CoordenadorBaseView, ListView):
+    template_name = 'listar/lista_professores.html'
     model = Professor
     login_url = reverse_lazy('login')
     context_object_name = 'professores'
@@ -276,54 +280,8 @@ class CoordenadorListaProfessoresView(CoordenadorView, ListView):
         messages.success(self.request, 'Professor deletado com sucesso!')
         return redirect('coordenador_lista_professores')
 
-class CoordenadorCreateProfessorView(CoordenadorView, TemplateView):
-    template_name = 'usuarios/coordenadores/cria_professor.html'
-
-    def post(self, *args, **kwargs):
-        nome = self.request.POST.get('nome')
-        cpf = self.request.POST.get('cpf')
-        telefone = self.request.POST.get('telefone')
-        email = self.request.POST.get('email')
-        username = self.request.POST.get('username')
-        senha = self.request.POST.get('senha')
-        coordenador = Coordenador.objects.get(user=self.request.user)
-        
-        context = {
-            'nome':nome, 'cpf':cpf, 'telefone':telefone, 'email':email, 'user':username, 'senha':senha}
-
-        #verificando se o username já existe para usuario
-        if User.objects.filter(username=username).exists():
-            messages.error(self.request, f'O usuário "{username}" não está disponível')
-            return render(self.request, self.template_name, context)
-        
-        #verificando o cpf
-        if Professor.objects.filter(cpf=cpf).exists():
-            messages.error(self.request, 'Um professor com esse CPF já está cadastrado')
-            return render(self.request, self.template_name, context)
-        
-        #criando user
-        user = User.objects.create_user(
-            username=username, 
-            email=email, 
-            password=senha, 
-            first_name=nome.split(' ')[0],
-            last_name=nome.split(' ')[-1]
-            )
-        user.save()
-
-        Professor.objects.create(
-            nome_completo=nome,
-            cpf=cpf,
-            telefone=telefone,
-            email=email,
-            coordenador=coordenador,
-            user=user
-        )
-        messages.success(self.request, 'Professor cadastrado com sucesso!')
-        return redirect('coordenador_lista_professores')
-    
-class CoordenadorEditProfessorView(CoordenadorView, DetailView):
-    template_name = 'usuarios/coordenadores/edita_professor.html'
+class CreateProfessorView(CoordenadorBaseView, DetailView):
+    template_name = 'editar/edita_professor.html'
     model = Professor
     context_object_name = 'professor'
 
@@ -357,8 +315,45 @@ class CoordenadorEditProfessorView(CoordenadorView, DetailView):
         
         return redirect('coordenador_lista_professores')
 
-class CoordenadorListaAlunosView(CoordenadorView, ListView):
-    template_name = 'usuarios/coordenadores/lista_alunos.html'
+class EditProfessorView(CoordenadorBaseView, DetailView):
+    template_name = 'editar/edita_professor.html'
+    model = Professor
+    context_object_name = 'professor'
+
+    def post(self, *args, **kwargs):
+        nome = self.request.POST.get('nome')
+        username = self.request.POST.get('username')
+        email = self.request.POST.get('email')
+        cpf = self.request.POST.get('cpf')
+        telefone = self.request.POST.get('telefone')
+        senha = self.request.POST.get('senha') or None
+        id = self.request.POST.get('id')
+
+        #update
+        professor = Professor.objects.get(id=id)
+        user = professor.user
+
+        user.first_name = nome.split(' ')[0]
+        user.last_name = nome.split(' ')[-1]
+        user.email = email
+        if senha:
+            user.set_password(senha)
+        user.save()
+
+        professor.nome_completo = nome
+        professor.email = email
+        professor.cpf = cpf
+        professor.telefone = telefone
+
+        professor.save()
+        messages.success(self.request, 'Professor editado com sucesso!')
+        
+        return redirect('coordenador_lista_professores')
+
+
+
+class ListaAlunosView(CoordenadorBaseView, ListView):
+    template_name = 'listar/lista_alunos.html'
     model = Aluno
     login_url = reverse_lazy('login')
     context_object_name = 'alunos'
@@ -385,8 +380,8 @@ class CoordenadorListaAlunosView(CoordenadorView, ListView):
         messages.success(self.request, 'Aluno deletado com sucesso!')
         return redirect('coordenador_lista_alunos', aluno.turma.id)
     
-class CoordenadorCreateAlunoView(CoordenadorView, TemplateView):
-    template_name = 'usuarios/coordenadores/cria_aluno.html'
+class CreateAlunoView(CoordenadorBaseView, TemplateView):
+    template_name = 'criar/cria_aluno.html'
     
     
 
@@ -399,8 +394,8 @@ class CoordenadorCreateAlunoView(CoordenadorView, TemplateView):
         messages.success(self.request, 'Aluno cadastrado com sucesso!')
         return redirect('coordenador_lista_alunos', turma.id)
     
-class CoordenadorEditAlunoView(CoordenadorView, DetailView):
-    template_name = 'usuarios/coordenadores/edita_aluno.html'
+class EditAlunoView(CoordenadorBaseView, DetailView):
+    template_name = 'editar/edita_aluno.html'
     model = Aluno
     context_object_name = 'aluno'
 
@@ -417,8 +412,9 @@ class CoordenadorEditAlunoView(CoordenadorView, DetailView):
         
         return redirect('coordenador_lista_alunos', turma.id)
 
-class CoordenadorListaTurmasView(CoordenadorView, ListView):
-    template_name = 'usuarios/coordenadores/lista_turmas.html'
+
+class ListaTurmasView(CoordenadorBaseView, ListView):
+    template_name = 'listar/lista_turmas.html'
     model = Turma
     login_url = reverse_lazy('login')
     context_object_name = 'turmas'
@@ -426,33 +422,36 @@ class CoordenadorListaTurmasView(CoordenadorView, ListView):
     ordering = ['nome']
 
     def get_queryset(self) -> QuerySet[Any]:
-        qs = super().get_queryset()
-        coordenador = Coordenador.objects.get(user=self.request.user)
-        escola = coordenador.escola
-        qs = qs.filter(escola=escola)
-        return qs
+        return super().get_queryset().filter(escola=self.request.user.coordenador.escola)
 
     def post(self, *args, **kwargs):
         id = self.request.POST.get('id')
-        turma = Turma.objects.get(id=id)
-        turma.delete()
-        messages.success(self.request, 'Turma deletada com sucesso!')
+        turma = Turma.objects.filter(id=id)
+        if turma.exists():
+            turma = turma[0]
+            turma.delete()
+            messages.success(self.request, 'Turma deletada com sucesso!')
+            return redirect('coordenador_lista_turmas')
+        messages.error(self.request, 'Turma não encontrada')
         return redirect('coordenador_lista_turmas')
 
-class CoordenadorCreateTurmaView(CoordenadorView, TemplateView):
-    template_name = 'usuarios/coordenadores/cria_turma.html'
+class CreateTurmaView(CoordenadorBaseView, TemplateView):
+    template_name = 'criar/cria_turma.html'
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['nome_escola'] = self.request.user.coordenador.escola.nome
+        return context
 
     def post(self, *args, **kwargs):
         nome = self.request.POST.get('nome')
-        coordenador = Coordenador.objects.get(user=self.request.user)
-        escola = coordenador.escola
+        grau = self.request.POST.get('grau')
 
-        Turma.objects.create(nome=nome, escola=escola)
+        Turma.objects.create(nome=nome, escola=self.request.user.coordenador.escola, grau=grau)
         messages.success(self.request, 'Turma cadastrada com sucesso!')
         return redirect('coordenador_lista_turmas')
     
-class CoordenadorEditTurmaView(CoordenadorView, DetailView):
-    template_name = 'usuarios/coordenadores/edita_turma.html'
+class EditTurmaView(CoordenadorBaseView, DetailView):
+    template_name = 'editar/edita_turma.html'
     model = Turma
     context_object_name = 'turma'
     def get_context_data(self, **kwargs):
@@ -471,6 +470,183 @@ class CoordenadorEditTurmaView(CoordenadorView, DetailView):
         messages.success(self.request, 'Turma editada com sucesso!')
         
         return redirect('coordenador_lista_turmas')
+
+class ListaAvaliacoesTurmaView(CoordenadorBaseView, DetailView):
+    template_name = 'usuarios/coordenadores/lista_avaliacoes_turma.html'
+    model = Turma
+    login_url = reverse_lazy('login')
+    context_object_name = 'turma'
+    paginate_by = 10
+    ordering = ['nome']
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        grau_turma = self.object.grau
+        avaliacoes = Simulado.objects.filter(grau_ensino=grau_turma)
+        respondidas = self.object.simulados_respondidos.all()
+        avaliacoes = [avaliacao for avaliacao in avaliacoes if avaliacao not in respondidas]
+        context['avaliacoes'] = avaliacoes
+
+        self.request.session['turma_id'] = self.object.id
+        return context
+
+class PreencherGabaritoView(CoordenadorBaseView, DetailView):
+    template_name = 'usuarios/coordenadores/preencher_gabarito_avaliacao.html'
+    model = Simulado
+    context_object_name = 'avaliacao'
+
+    def get_context_data(self, **kwargs):
+        if self.request.method == 'GET':
+            context = super().get_context_data(**kwargs)
+            simulado = self.object
+            self.request.session['simulado_id'] = simulado.id
+            questoes = QuestaoReferencia.objects.filter(simulado=simulado).order_by('numero_questao')
+            context['questoes'] = questoes
+            context['simulado'] = simulado
+            context["turma"] = Turma.objects.get(id=self.request.session.get('turma_id')) 
+            context["alunos"] = Aluno.objects.filter(turma=context["turma"])
+            return context
+        return super().get_context_data(**kwargs)
+
+    def post(self, *args, **kwargs):
+        #pegando todas as respostas com resposta_questao_
+        for key, value in self.request.POST.items():
+            if 'resposta_questao_' in key:
+                aluno_id = key.split('_')[-1]
+                questao_id = key.split('_')[-2]
+                questao = QuestaoReferencia.objects.get(id=questao_id)
+                alternativa_correta = questao.questao.alternativa_correta
+                alternativa_selecionada = value
+                acertou = True if alternativa_correta == alternativa_selecionada else False
+                Resposta.objects.create(
+                    questao_referencia=questao,
+                    resposta=alternativa_selecionada,
+                    aluno=Aluno.objects.get(id=aluno_id),
+                    acertou=acertou
+                )
+                if acertou:
+                    if not questao.quantidade_respostas_corretas:
+                        questao.quantidade_respostas_corretas = 0
+                    questao.quantidade_respostas_corretas += 1
+
+                if alternativa_selecionada == 'a':
+                    if not questao.quantidade_respostas_a:
+                        questao.quantidade_respostas_a = 0
+                    questao.quantidade_respostas_a += 1
+                elif alternativa_selecionada == 'b':
+                    if not questao.quantidade_respostas_b:
+                        questao.quantidade_respostas_b = 0
+                    questao.quantidade_respostas_b += 1
+                elif alternativa_selecionada == 'c':
+                    if not questao.quantidade_respostas_c:
+                        questao.quantidade_respostas_c = 0
+                    questao.quantidade_respostas_c += 1
+                elif alternativa_selecionada == 'd':
+                    if not questao.quantidade_respostas_d:
+                        questao.quantidade_respostas_d = 0
+                    questao.quantidade_respostas_d += 1
+
+                questao.save()
+
+        #adicionando o simulado a turma
+        turma = Turma.objects.get(id=self.request.session.get('turma_id'))
+        turma.simulados_respondidos.add(Simulado.objects.get(id=self.request.session.get('simulado_id')))
+        turma.save()
+        messages.success(self.request, 'Gabarito preenchido com sucesso!')
+        return redirect('coordenador_lista_turmas')
+
+
+############# COORDENADOR GERAL ################
+class CoordenadorGeralBaseView(LoginRequiredMixin):
+    login_url = reverse_lazy('login')
+
+    #criando redirect para deslogar caso o usuario não seja coordenador
+    def dispatch(self, request, *args, **kwargs):
+        if not CoordenadorGeral.objects.filter(user=request.user).exists():
+            logout(request)
+            return redirect('login')
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        user = self.request.user
+        self.request.session['nome'] = self.request.user.first_name
+        self.request.session['classe'] = "Coordenador Geral"
+        self.request.session["coordenador_geral_id"] = CoordenadorGeral.objects.get(user=user).id
+        return context
+
+class DashboardCoordenadorGeral(CoordenadorGeralBaseView, TemplateView):
+    template_name = 'usuarios/coordenadores_gerais/dashboard.html'
+
+class ListaAvaliacoesView(CoordenadorGeralBaseView, ListView):
+    template_name = 'listar/lista_avaliacao.html'
+    model = Simulado
+    login_url = reverse_lazy('login')
+    context_object_name = 'avaliacoes'
+    paginate_by = 10
+    ordering = ['nome']
+
+    def get_queryset(self) -> QuerySet[Any]:
+        qs = super().get_queryset()
+        qs = qs.filter(responsavel=self.request.user)
+        return qs
+
+    def post(self, *args, **kwargs):
+        id = self.request.POST.get('id')
+        simulado = Simulado.objects.get(id=id)
+        simulado.delete()
+        messages.success(self.request, 'Simulado deletado com sucesso!')
+        return redirect('coordenador_lista_avaliacoes')
+
+class CreateAvaliacaoView(CoordenadorGeralBaseView, TemplateView):
+    template_name = 'criar/cria_avaliacao.html'
+
+    def post(self, *args, **kwargs):
+        simulado = Simulado.objects.create(
+            nome=self.request.POST.get('nome'),
+            responsavel=self.request.user,
+            tipo_simulado="questoes_construidas",
+            grau_ensino=self.request.POST.get('grau_ensino'),
+            matriz_referencial=self.request.POST.get('matriz_referencial'),
+        )
+
+        questoes = self.request.POST.get('questoes')
+        if questoes:
+            questoes = eval(questoes)
+        
+        for numero_questao, questao in enumerate(questoes):
+            questao = Questao.objects.create(
+                unidade_tematica_texto=questao["unidade_tematica"],
+                habilidades_abncc_texto=questao["habilidades_abncc"],
+                descritor=questao["descritor"],
+                alternativa_correta=questao["alternativa_correta"],
+                componente=questao["componente"],
+            )
+
+
+
+            QuestaoReferencia.objects.create(
+                simulado = simulado,
+                numero_questao = numero_questao+1,
+                questao = questao,
+            )
+
+        messages.success(self.request, 'Avaliação cadastrada com sucesso!')
+        return redirect('coordenador_lista_avaliacoes')
+
+class VisualizarAvaliacaoView(CoordenadorGeralBaseView, DetailView):
+    template_name = 'visualizar/ver_avaliacao.html'
+    model = Simulado
+    context_object_name = 'avaliacao'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        simulado = Simulado.objects.get(id=self.kwargs.get('pk'))
+        questoes = QuestaoReferencia.objects.filter(simulado=simulado)
+        context['questoes'] = questoes
+        context['simulado'] = simulado
+        return context
+
 ############# PROFESSOR ################
 class ProfessorView(LoginRequiredMixin):
     login_url = reverse_lazy('login')
@@ -535,7 +711,6 @@ class ProfessorListView(AlunoView, ListView):
             turmas_g["turma"] = turma
             turmas_alunos.append(turmas_g)
         context['turmas_alunos'] = turmas_alunos
-        print(context['turmas_alunos'][0].get('alunos'))    
         return context
 
     def post(self, *args, **kwargs):
